chore: Update header.php with Swiper CSS and JS dependencies
The header.php file has been updated to include the necessary CSS and JS dependencies for the Swiper library. This will enable the use of Swiper's slider functionality on the website. Additionally, some custom styling has been added to the Swiper wrapper and pagination bullets to enhance the visual appearance of the slider.